refactor(smb): move metadata write/enrich cluster to SmbMetadata
Move writeMetadata, writeMetadataWithDetail, enrichWithGalleryTags,
writeMetadataSkeleton and enrichLocalMetadataIfMissing out of SmbStorage into
SmbMetadata, which now owns metadata.json writing end to end. The methods
resolve the gallery dir via the now package-private SmbStorage.getGalleryDir
and reference SmbStorage.METADATA_FILE / isConfigured.

Callers updated: SmbStorage.syncDownloadedGallery + finalizeDownloadedGallery
call SmbMetadata.writeMetadataWithDetail; SmbAutoDownloadManager and
LocalInventoryScene call the moved skeleton/enrich entry points. Drops the
now-unused EhEngine, EhUrl, GalleryDetail, GalleryTagGroup and
IoThreadPoolExecutor imports from SmbStorage.
